Use uint32 for Nlink, matching fuse.
Cf. https://github.com/torvalds/linux/blob/00c570f/include/uapi/linux/fuse.h#L159

+++ fuseops/convert.go
@@ -472,7 +472,7 @@ func convertAttributes(
 	out.Mtime, out.MtimeNsec = convertTime(in.Mtime)
 	out.Ctime, out.CtimeNsec = convertTime(in.Ctime)
 	out.SetCrtime(convertTime(in.Crtime))
-	out.Nlink = uint32(in.Nlink) // TODO(jacobsa): Make the public field uint32?
+	out.Nlink = in.Nlink
 	out.Uid = in.Uid
 	out.Gid = in.Gid
